--- database/users.txt
@@ -1,2 +1,0 @@
-Username: admin
-Password: 1234
